requirements.txt
# Hermes Miner - NL2GraphQL for Bittensor Subnet 82

# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
httpx>=0.25.0

# GraphQL
graphql-core>=3.2.0

# LLM Client - Multi-provider support
openai>=1.0.0

# LangChain & LangGraph (for ReAct agent)
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Vector Database & Embeddings (optional, for RAG)
weaviate-client>=4.0.0
sentence-transformers>=2.2.0

# Machine Learning (optional, for fine-tuning)
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.6.0
trl>=0.7.0

# Unsloth for fine-tuning (optional - requires GPU)
# unsloth @ git+https://github.com/unslothai/unsloth.git

# vLLM for inference (optional - requires GPU)
# vllm>=0.6.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
